third_party/blink/web_tests/http/tests/security/contentSecurityPolicy/xsl-img-blocked-expected.txt

CONSOLE WARNING: XSLTProcessor and XSLT Processing Instructions have been deprecated by all browsers. These features will be removed from this browser soon. See https://chromestatus.com/feature/4709671889534976.
CONSOLE ERROR: Loading the image 'http://127.0.0.1:8000/security/resources/abe.png' violates the following Content Security Policy directive: "img-src 'none'". The action has been blocked.
Here is an image: 
